sqlsmith: make the generic file not a .go file
