+++ pkg/internal/sqlsmith/Makefile
@@ -1,5 +1,5 @@
 # Use FORCE here because we want to regenerate when this makefile changes.
 gen-sampler-gen.go: FORCE
-	genny -in=sampler.go -out=gen-sampler.go gen "element=statement,tableExpr,selectStatement,scalarExpr"
-	../../../bin/crlfmt -w gen-sampler.go >/dev/null
+	genny -in=sampler.tmpl -out=sampler.go gen "element=statement,tableExpr,selectStatement,scalarExpr"
+	../../../bin/crlfmt -w sampler.go >/dev/null
 FORCE:
